LayoutTests/platform/glib/imported/w3c/web-platform-tests/xhr/send-authentication-basic-cors-expected.txt

CONSOLE MESSAGE: Preflight response is not successful. Status code: 404
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.web-platform.test:8800/xhr/resources/auth1/corsenabled.py due to access control checks.

PASS XMLHttpRequest: send() - "Basic" authenticated CORS requests with user name and password passed to open() (asserts failure)
